issue124: extract Change and implementing classes
