+++ javaparser-core/src/main/java/com/github/javaparser/printer/lexicalpreservation/changes/Change.java
@@ -1,0 +1,9 @@
+package com.github.javaparser.printer.lexicalpreservation.changes;
+
+import com.github.javaparser.ast.Node;
+import com.github.javaparser.printer.concretesyntaxmodel.CsmConditional;
+
+public interface Change {
+
+    boolean evaluate(CsmConditional csmConditional, Node node);
+}
